Update paris path  data test

+++ data/tests/utilities.py
@@ -10,7 +10,7 @@ def prep_dict_for_compare(dict):
         
     # If dict values are floats, add acceptible relative tolerance of 1e-8
     # See pytest.approx: https://docs.pytest.org/en/7.1.x/reference/reference.html#pytest-approx
-    if all([isinstance(v, numpy.float64) for v in dict.values()]):
+    if all([isinstance(v, float) for v in dict.values()]):
         dict = {i:pt.approx(x, rel=1e-8) for i, x in dict.items()}
         
     return dict
